==> Current/SECOND_TRIAL/Misc/03_svm_model.ipynb <==
import pandas as pd
import numpy as np
import ast

from nrclex import NRCLex
from sklearn import utils
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.linear_model import LogisticRegression
from sklearn import model_selection, naive_bayes, svm
from sklearn.metrics import accuracy_score
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import classification_report

from nltk.corpus import stopwords
stopwords = stopwords.words('english')

import nltk
nltk.download('vader_lexicon')
from nltk.sentiment.vader import SentimentIntensityAnalyzer

sid = SentimentIntensityAnalyzer()

df = pd.read_csv("./Dataset Creation/well-narratives.csv")

# Cleaning

def clean_text(df, column_name):
    df['cleaned_text'] = df[column_name].fillna('')
    df['cleaned_text'] = df['cleaned_text'].str.lower()
    df['cleaned_text'] = df['cleaned_text'].str.replace(r'(@[A-Za-z0-9]+)|([^0-9A-Za-z \t])|(\w+:\/\/\S+)|rt|\d+', '')
    df['cleaned_text'] = df['cleaned_text'].str.replace(r'^\s+|\s+$', '') 
    df['cleaned_text'] = df['cleaned_text'].apply(lambda x: ' '.join([w for w in x.split() if w not in (stopwords)]))
    return df

df_cleaned = clean_text(df, "sentence")

# Tagging

def tag_sentiment(string):
    result = sid.polarity_scores(string)
    if (result['compound'] > 0):
        return "POS"
    elif (result['compound'] == 0):
        return "NEU"
    else:
        return "NEG"

df_cleaned["sentiment_tag"] = df_cleaned["cleaned_text"].apply(lambda x : tag_sentiment(x))

# Results

df_cleaned['sentiment_tag'].value_counts()

df.head()

# Filtering

def filter_sentiment(row):
    text_object = NRCLex(row["cleaned_text"])
    return str(text_object.affect_frequencies)

df_cleaned["sentiments"] = df_cleaned.apply(filter_sentiment, axis = 1)

df_cleaned.head()

# SVM Model

sentiments = pd.DataFrame(columns=['cleaned_text', 'tag', 'fear', 'anger', 'anticipation', 'trust', 'surprise', 
                                   'positive', 'negative', 'sadness', 'disgust', 'joy'])

for index, row in df_cleaned.iterrows():
    narrative = row['cleaned_text']
    tag = row['sentiment_tag']
    breakdown = ast.literal_eval(row['sentiments'])
    
    breakdown['cleaned_text'] = narrative
    breakdown['tag'] = tag
    sentiments = sentiments.append(breakdown, ignore_index=True)

sentiments = sentiments.replace(np.nan, 0)

Train_X, Test_X, Train_Y, Test_Y = model_selection.train_test_split(sentiments[['fear', 'anger', 'anticipation', 'trust', 'surprise', 
                                   'positive', 'negative', 'sadness', 'disgust', 'joy']],sentiments['tag'],test_size=0.3,random_state=42)

SVM = svm.SVC(C=1.0, kernel='linear', degree=3, gamma='auto')
SVM.fit(Train_X,Train_Y)

predictions_SVM = SVM.predict(Test_X)
print("SVM Accuracy Score: ",accuracy_score(predictions_SVM, Test_Y)*100)

print(classification_report(Test_Y, predictions_SVM))

# Naive Bayes

Naive = naive_bayes.MultinomialNB()
Naive.fit(Train_X,Train_Y)

predictions_NB = Naive.predict(Test_X)

print("Naive Bayes Accuracy Score: ", accuracy_score(predictions_NB, Test_Y)*100)

print(classification_report(Test_Y, predictions_NB))
