==> Current/SECOND_TRIAL/Misc/03_text_analysis.ipynb <==
# Text analysis

Hello everyone! We're learning about text analysis in the section with `TextBlob`, `Scikit-learn`, and `WordCloud`. We'll take a look into what our tweets contain and how we can visualize them to get a better understanding of their makeup.

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.style as style
from wordcloud import WordCloud
from textblob import TextBlob as tb
from IPython.display import Image as im
style.use('fivethirtyeight')

# Read cleaned tweets, term frequencies of DTM, ngram, and TFIDF
freq_dtm = pd.read_csv('out/frequency_dtm.csv', index_col=None, header=0)
freq_ngram = pd.read_csv('out/frequency_ngram.csv', index_col=None, header=0)
freq_tfidf = pd.read_csv('out/frequency_tfidf.csv', index_col=None, header=0)
cleaned_tweets = pd.read_csv('out/cleaned_tweets.csv', index_col=None, header=0)
cleaned_tweets.head()

### Visualization of term frequencies

# Visualize frequencies
sns.barplot(data=freq_dtm.head(10), x='term',
            y='frequency').set_title('Frequent terms in DTM')

# Visualize frequencies
sns.barplot(data=freq_ngram.head(10), x='term',
            y='frequency').set_title('Frequent terms in Ngram')

# Visualize frequencies
sns.barplot(data=freq_tfidf.head(10), x='term',
            y='frequency').set_title('Frequent terms in TFIDF')

### Sentiment analysis

def analyze_tweets(tweets):
    """Analyzes the sentiment polarity and subjectivity of tweets"""
    sentiment = []
    subjectivity = []
    for text in tweets['text']:
        blob = tb(text)
        sentiment.append(blob.polarity)
        subjectivity.append(blob.subjectivity)
    tweets['sentiment'] = sentiment
    tweets['subjectivity'] = subjectivity
    return tweets

# Analyze the sentiment and subjectivity of tweets
analyzed_tweets = analyze_tweets(cleaned_tweets)
analyzed_tweets.head()

# Visualize distribution of sentiment and subjectivity
sns.lmplot(data=analyzed_tweets, x='sentiment', y='subjectivity', fit_reg=True) 

# Get one tweet analysis
val = analyzed_tweets.iloc[1]
print('Original text:', val['text'])
print('Cleaned text:', val['cleaned_text'])
print('Sentiment:', val['sentiment'])
print('Subjectivity:', val['subjectivity'])

### Wordcloud

def create_wordcloud(tweets, max_words=500):
    """Create a wordcloud of most common words in a set of tweets"""
    
    # Transform text for WordCloud
    tweets = tweets['cleaned_text']
    tweets = tweets.dropna()
    tweets = ' '.join(tweets)
    tweets = tweets.replace(' ', ',')
    
    # Generate wordcloud image
    wc = WordCloud(background_color="white", max_words=max_words, colormap='plasma')
    wc.generate(tweets)
    plt.imshow(wc, interpolation='bilinear')
    plt.title('Twitter Generated Cloud', size=30)
    plt.axis("off")

# Generate a wordcloud
create_wordcloud(cleaned_tweets)
