==> Current/SECOND_TRIAL/Misc/04_text_modeling.ipynb <==
# Text modeling

Hello everyone! We're going ahead and creating models of our tweets data using `GenSim`. `GenSim` provides facilities for topic modeling and text classification that we will leverage later on.

import pickle
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.style as style
import pyLDAvis, pyLDAvis.sklearn
from sklearn.manifold import TSNE
from sklearn.cluster import KMeans
from sklearn.decomposition import LatentDirichletAllocation, PCA
pyLDAvis.enable_notebook()
style.use('fivethirtyeight')

### Data preparation
We're reusing our generated tweets from the previous section. 

# Read cleaned texts
cleaned_tweets = pd.read_csv('out/cleaned_tweets.csv', index_col=None, header=0)
texts = cleaned_tweets['cleaned_text'].astype('U')
cleaned_tweets.head()

# Load existing vectorizers
dtm_v = pickle.load(open('out/dtm.pk', 'rb'))
ngram_v = pickle.load(open('out/ngram.pk', 'rb'))
tfidf_v = pickle.load(open('out/tfidf.pk', 'rb'))

# Load terms from existing vectorizers
dtm_terms = dtm_v.get_feature_names()
ngram_terms = ngram_v.get_feature_names()
tfidf_terms = tfidf_v.get_feature_names()

# Transform texts into representations
dtm = dtm_v.fit_transform(texts)
ngram = ngram_v.fit_transform(texts)
tfidf = tfidf_v.fit_transform(texts)

### Latent Dirichlet Allocation
`LDA` is used for `topic modeling` by classifying text in a document to a particular topic. It builds a topic per document model and words per topic model, modeled as Dirichlet distributions. A lot of math, but the important thing to remember is it can determine related terms that form `topics`.

We are going to user an additional package `pyLDAvis` to visualize our LDA results

def lda(texts):
    """
    Executes LDA algorithm to determine topics of the corpus
    """
    lda_model = LatentDirichletAllocation(
        n_components=5, max_iter=10,
        learning_method='online', 
        learning_offset=50., random_state=0)
    lda_model.fit_transform(texts)
    return lda_model

# Execute LDA
dtm_lda = lda(dtm)
ngram_lda = lda(ngram)
tfidf_lda = lda(tfidf)

pyLDAvis.sklearn.prepare(dtm_lda, dtm, dtm_v)

pyLDAvis.sklearn.prepare(ngram_lda, ngram, ngram_v)

pyLDAvis.sklearn.prepare(tfidf_lda, tfidf, tfidf_v)

### K-means clustering
We can find interesting groupings that come out of the data using `K-means`. This algorithm creates `clusters` of data by separating samples into `n` number of groups of equal variance that surround centroids. 

Because our text data representations have many dimensions (1000, 2000), it will be difficult to visualize this properly. However, we will be using dimensionality reduction techniques such as `Prinicipal Component Analysis` and `t-Distributed Stochastic Neighbor Embedding` to reduce our dimensions to two. Again, a lot of mathematics, but an important thing to remember is that both "combine" existing dimensions until only the desired number of dimensions is reached. 

def kmeans(texts, clusters=5):
    """
    Execute Kmeans algorithm to generate clusters
    """
    kmeans_model = KMeans(n_clusters=clusters, max_iter=1000)
    kmeans_model.fit(texts)
    kmeans_result = kmeans_model.predict(texts)
    return kmeans_result

def plot_kmeans(texts, clusters): 
    """
    Generates visualizations of clusters using Prinicipal Component Analysis
    and t-Distributed Stochastic Neighbor Embedding
    """
    pca = PCA(n_components=2).fit_transform(texts.todense())
    tsne = TSNE().fit_transform(PCA(n_components=50).fit_transform(texts.todense()))
    pca_a = pca[:, 0]
    pca_b = pca[:, 1]
    tsne_a = tsne[:, 0]
    tsne_b = tsne[:, 1]
    
    f, ax = plt.subplots(1, 2)
    sns.scatterplot(x=pca_a, y=pca_b, hue=clusters, ax=ax[0], )
    sns.scatterplot(x=tsne_a, y=tsne_b, hue=clusters, ax=ax[1], )

dtm_kmeans = kmeans(dtm)
plot_kmeans(dtm, dtm_kmeans)

ngram_kmeans = kmeans(ngram)
plot_kmeans(ngram, ngram_kmeans)

tfidf_kmeans = kmeans(tfidf)
plot_kmeans(tfidf, tfidf_kmeans)
